+++ src/date_func.h
@@ -17,7 +17,7 @@ extern Date      _date;
 extern DateFract _date_fract;
 extern uint64    _tick_counter;
 extern uint8     _tick_skip_counter;
-extern uint32    _scaled_tick_counter;
+extern uint64    _scaled_tick_counter;
 extern DateTicksScaled _scaled_date_ticks;
 extern uint32    _quit_after_days;
